style: rename ShellSortVisualisation directory to ShellSortVisualization for consistency
