+++ src/theme/MDXComponents.js
@@ -7,7 +7,7 @@ import InsertionSortVisualization from "@site/src/components/DSA/arrays/Insertio
 import QuickSortVisualization from "@site/src/components/DSA/arrays/QuickSortVisualization";
 import HeapSortVisualization from "@site/src/components/DSA/arrays/HeapSortVisualization";
 import MergeSortVisualization from "@site/src/components/DSA/arrays/MergeSortVisualization";
-import ShellSortVisualization from "@site/src/components/DSA/arrays/ShellSortVisualisation";
+import ShellSortVisualization from "@site/src/components/DSA/arrays/ShellSortVisualization";
 import SelectionSortVisualization from "@site/src/components/DSA/arrays/SelectionSortVisualization";
 import DijkstraVisuzalizations from "@site/src/components/DSA/graphs/DijkstraVisualizations";
 import FloydWarshallVisualizations from "@site/src/components/DSA/graphs/FloydWarshallVisualizations";
